Added CBC.py and channel.py layers to separate the config dictionaries from the functions that manipulate them
