+++ RP API/CBC_config.py
@@ -19,6 +19,8 @@ CBC_keys = ["CBC_enabled",
             "velocity_external",
             "displacement_external",
             "polynomial_target",
+            "kp",
+            "kd",
             "r_hat_start",
             "r_hat_stop",
             "r_hat_sweep",
@@ -50,6 +52,8 @@ datatypes = {"CBC_enabled": bool,
              "velocity_external": bool,
              "displacement_external": bool,
              "polynomial_target": str,
+             "kp": float,
+             "kd": float,
              "r_hat_start": float,
              "r_hat_stop": float,
              "r_hat_sweep": bool,
@@ -76,6 +80,8 @@ limits = {"CBC_enabled": [0, 1],
           "velocity_external": [0, 1],
           "displacement_external": [0, 1],
           "polynomial_target": ["displacement", "velocity"],
+          "kp": [-1024,1024],
+          "kd": [-1024,1024],
           "r_hat_start": [-1000, 1000],
           "r_hat_stop": [-1000, 1000],
           "r_hat_sweep": [0, 1],
@@ -123,8 +129,8 @@ class CBC_config(dict):
         value checks before setting any items. """
 
         if key in CBC_keys:
-            self.choose_external_input_type(key, value)    
-    
+            self.choose_external_input_type(key, value)
+
             if type(value) == datatypes[key]:
                 if self.is_within_limits(value, limits[key]):
                     super().__setitem__(key, value)
@@ -153,114 +159,3 @@ class CBC_config(dict):
                 raise ValueError("Limits list must contain exactly two items.")
             lower_limit, upper_limit = sorted(limits)
             return lower_limit <= value <= upper_limit
-    
-    def choose_external_input_type(self, key, value):
-        """
-        This function configures whether the input to the RP is of a certain type.
-        In doing so, it sets its compliment to False. 
-        
-        
-        --- Logic ---    
-        (input) key = value
-            -> displacement_external = ...  (output)
-            -> displacement_external = ...  (output)
-        
-        displacement_external = False
-            -> displacement_external = False
-            -> velocity_external = False
-            
-        velocity_external = False
-            -> displacement_external = False
-            -> velocity_external = False
-            
-        displacement_external = True
-            -> displacement_external = True
-            -> velocity_external = False
-            
-        velocity_external = True
-            -> displacement_external = False
-            -> velocity_external = True
-        """
-        if key == "velocity_external" and value == True:
-            self["displacement_external"] = False
-            print("Warning: key '*displacement_external' set to 'False'.")
-        elif key == "displacement_external" and value == True:
-            self["velocity_external"] = False
-            print("Warning: key '*velocity_external' set to 'False'.")
-    
-    def set_param(self, param_name, param_val):
-        """
-        Sets relevant parameter value. 
-        
-        If the parameter can be swept (such as frequency), a list or tuple can 
-        be given to set the sweep range. As such, the sweep logic is simultaneously 
-        updated.
-        """
-        if param_name in CBC_sweepable:
-            set_value_or_sweep(self, param_name, param_val, CBC_sweepable)
-        else:
-            self[param_name] = param_val
-            
-    def set_params_CBC(self, **kwargs):
-        """
-        Sets the parameters for the CBC output type
-        Coefficients are of the mathematical form of:
-            y = Ax + B
-        
-        Possible key/arguments
-            1) A - int, float, list, tuple
-            2) B - int, float, list, tuple
-            3) C - int, float, list, tuple
-            4) D - int, float, list, tuple
-            5) rhat - int, float, list, tuple
-            6) frequency - int, float, list, tuple
-            7) polynomial_target - "displacement" or "velocity"
-            8) external - "displacement" or "velocity"
-            9) input_order - 1, 2
-    
-        Empty/non-assigned arguments will be ignored/left unchanged.
-        Irrelevant arguments outside of the possible keys will be ignored.
-        
-        
-        Usage:
-            RP.CBC.set_params_CBC(A=1, B=[1,4])
-                -> A_start = 1
-                -> A_stop = 0
-                -> A_sweep = False
-                -> B_start = 1
-                -> B_stop = 4
-                -> B_sweep = True
-                -> All other parameters are ignored.
-            
-            RP.CBC.set_params_CBC(polynomial_target="displacement")
-                -> polynomial_target = "displacement"
-                -> All other parameters are ignored
-            
-            RP.CBC.set_params_CBC(external="displacement")
-                -> displacement_external = True
-                -> velocity_external = False
-                -> All other parameters are ignored
-        """        
-        CBC_keys = {"A", "B", "C", "D", "rhat", "frequency", "polynomial_target", "external", "input_order"}
-        for key, value in kwargs.items():
-            if key == "external":
-                if value == "displacement":
-                    self.set_param("displacement_external", True)
-                    CBC_keys.remove(key)
-                elif value == "velocity":
-                    self.set_param("velocity_external", True)
-                    CBC_keys.remove(key)
-                else:
-                    print("Warning: value '%s' is not valid for key '%s', and will be ignored." % (value, key))
-            elif key in CBC_keys:
-                self.set_param(key, value)
-                CBC_keys.remove(key)
-            else:
-                # TODO: Could throw a KeyError(?) telling that this key is not used for the linear mode
-                print("Warning: key '%s' is not used in CBC, and will be ignored." % key)
-        
-        for key in CBC_keys:
-            print("Warning: key '%s' not found. Value will remain unchanged" % key)
-
-
-
